+++ ktrain/graph/predictor.py
@@ -5,7 +5,9 @@ from .. import utils as U
 
 class NodePredictor(Predictor):
     """
+    ```
     predicts graph node's classes
+    ```
     """
 
     def __init__(self, model, preproc, batch_size=U.DEFAULT_BS):
@@ -30,8 +32,10 @@ class NodePredictor(Predictor):
 
     def predict_transductive(self, node_ids, return_proba=False):
         """
+        ```
         Performs transductive inference.
         If return_proba is True, returns probabilities of each class.
+        ```
         """
         gen = self.preproc.preprocess_valid(node_ids)
         gen.batch_size = self.batch_size
@@ -44,8 +48,10 @@ class NodePredictor(Predictor):
 
     def predict_inductive(self, df, G, return_proba=False):
         """
+        ```
         Performs inductive inference.
         If return_proba is True, returns probabilities of each class.
+        ```
         """
 
         gen = self.preproc.preprocess(df, G)
@@ -59,7 +65,9 @@ class NodePredictor(Predictor):
 
 class LinkPredictor(Predictor):
     """
+    ```
     predicts graph node's classes
+    ```
     """
 
     def __init__(self, model, preproc, batch_size=U.DEFAULT_BS):
@@ -80,8 +88,10 @@ class LinkPredictor(Predictor):
 
     def predict(self, G, edge_ids, return_proba=False):
         """
+        ```
         Performs link prediction
         If return_proba is True, returns probabilities of each class.
+        ```
         """
         gen = self.preproc.preprocess(G, edge_ids)
         gen.batch_size = self.batch_size