+++ ktrain/tabular/data.py
@@ -51,7 +51,9 @@ def tabular_from_df(train_df, label_columns=[], date_columns=[], val_df=None, va
 def tabular_from_csv(train_csv, label_columns=[], date_columns=[], val_csv=None, val_pct=0.1, 
                      index_col=None, is_regression=False, max_card=20, random_state=None):
     """
+    ```
     Loads tabular data from CSV file
+    ```
     """
 
     # read in dataset