+++ ktrain/utils.py
@@ -66,6 +66,13 @@ def metrics_from_model(model):
             except:
                 warnings.warn(msg)
                 return []
+        elif isinstance(mlist, list) and hasattr(mlist[0], 'name'): # tf.keras.metrics.AUC()
+            try:
+                return [m.name for m in mlist]
+            except:
+                warnings.warn(msg)
+                return []
+
         else:
             warnings.warn(msg)
             return []
@@ -97,6 +104,8 @@ def is_classifier(model):
     else:
         mlist = metrics_from_model(model)
         if isinstance(mlist, (list, np.ndarray)) and any(['accuracy' in m for m in mlist]):
+            is_classifier = True
+        elif isinstance(mlist, (list, np.ndarray)) and any(['auc' in m for m in mlist]):
             is_classifier = True
 
     # check for multilabel
@@ -535,6 +544,7 @@ def get_hf_model_name(model_id):
 class YTransform:
     def __init__(self, class_names=[], label_encoder=None):
         """
+        ```
         Cheks and transforms array of targets. Targets are transformed in place.
         Args:
           class_names(list):  labels associated with targets (e.g., ['negative', 'positive'])
@@ -546,6 +556,7 @@ class YTransform:
                          2. targets are strings and task is classification (class_names are populated automatically)
           label_encoder(LabelEncoder): a prior instance of LabelEncoder.  
                                        If None, will be created when train=True
+        ```
         """
         if type(class_names) != list:
             if isinstance(class_names, (pd.Series, np.ndarray)): class_names = class_names.tolist()
@@ -621,11 +632,13 @@ class YTransform:
 class YTransformDataFrame(YTransform):
     def __init__(self, label_columns=[], is_regression=False):
         """
+        ```
         Checks and transforms label columns in DataFrame. DataFrame is modified in place
         Args:
           label_columns(list): list of columns storing labels 
           is_regression(bool): If True, task is regression and integer targets are treated as numeric dependent variable.
                                IF False, task is classification and integer targets are treated as class IDs.
+        ```
         """
         self.is_regression = is_regression
         if isinstance(label_columns, str): label_columns = [label_columns]